LayoutTests/imported/w3c/web-platform-tests/css/css-values/if-style-invalidation-expected.txt


FAIL CSS Values and Units Test: if() style() condition invalidation Error: assert_equals: expected "true_value" but got "if(style(--x: 3): true_value; else: false_value;)"
